+++ Module6/src/main/java/com/epam/training/gen/ai/controller/DocumentController.java
@@ -1,6 +1,5 @@
 package com.epam.training.gen.ai.controller;
 
-import com.epam.training.gen.ai.service.StoreService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -8,18 +7,20 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.util.function.Consumer;
+
 @RestController
 @RequestMapping("/api/v1/documents")
 public class DocumentController {
-    private final StoreService storeService;
+    private final Consumer<MultipartFile> storeService;
 
-    public DocumentController(StoreService storeService) {
+    public DocumentController(Consumer<MultipartFile> storeService) {
         this.storeService = storeService;
     }
 
     @PostMapping
     public ResponseEntity<Void> uploadMultipleFiles(@RequestParam("file") MultipartFile file) {
-        storeService.save(file);
+        storeService.accept(file);
         return ResponseEntity.noContent().build();
     }
 }
